requirements.txt
et-xmlfile==1.1.0
lxml==5.1.0
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.1
python-dateutil==2.9.0.post0
python-docx==1.1.0
pytz==2024.1
six==1.16.0
typing_extensions==4.10.0
tzdata==2024.1
